Mixed things
This is a commit that im not proud of. It includes a lot of new things.
Some new and reworked views. Some more components. Moved out the routes from app.jsx into a seperate file.

The main reason for this is because I have worked with the setup of the whole flow in the application. So now all of the views are accessible through the UI.